v0.3

+++ week06/main.go
@@ -8,6 +8,7 @@ import (
 func main() {
 	var i int = 13
 	f := 12.9
-	fmt.Printf("%d* %f=%f\n", i, f, float64(i)*f)
+	// fmt.Printf("%d* %f=%f\n", i, f, float64(i)*f)
+	fmt.Printf("%d* %f=%d\n", i, f, i*int(f))
 	fmt.Println(reflect.TypeOf(f))
 }
